+++ src/index.js
@@ -1,4 +1,0 @@
-"use strict";
-module.exports = require("./arangojs").default;
-module.exports.default = module.exports;
-Object.defineProperty(module.exports, "__esModule", { value: true });