+++ src/test/06-managing-functions.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { Database } from "../arangojs";
+import { Database } from "../database";
 
 const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
 const ARANGO_VERSION = Number(