+++ src/test/12-simple-queries.ts
@@ -1,5 +1,7 @@
 import { expect } from "chai";
-import { ArrayCursor, Database, DocumentCollection } from "../arangojs";
+import { DocumentCollection } from "../collection";
+import { ArrayCursor } from "../cursor";
+import { Database } from "../database";
 
 const range = (n: number): number[] => Array.from(Array(n).keys());
 const alpha = (i: number): string => String.fromCharCode("a".charCodeAt(0) + i);