+++ src/test/13-bulk-imports.ts
@@ -1,9 +1,6 @@
 import { expect } from "chai";
-import {
-  CollectionImportOptions,
-  Database,
-  DocumentCollection
-} from "../arangojs";
+import { CollectionImportOptions, DocumentCollection } from "../collection";
+import { Database } from "../database";
 
 const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
 const ARANGO_VERSION = Number(