+++ src/test/27-query-management.ts
@@ -1,5 +1,8 @@
 import { expect } from "chai";
-import { aql, ArangoError, ArrayCursor, Database } from "../arangojs";
+import { aql } from "../aql-query";
+import { ArrayCursor } from "../cursor";
+import { Database } from "../database";
+import { ArangoError } from "../error";
 
 const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
 const ARANGO_VERSION = Number(