+++ src/util/codes.ts
@@ -1,4 +1,11 @@
+/** @hidden */
 export const DOCUMENT_NOT_FOUND = 1202;
+
+/** @hidden */
 export const COLLECTION_NOT_FOUND = 1203;
+
+/** @hidden */
 export const DATABASE_NOT_FOUND = 1228;
+
+/** @hidden */
 export const GRAPH_NOT_FOUND = 1924;