+++ src/util/joinPath.ts
@@ -1,2 +1,4 @@
 import { posix } from "path";
+
+/** @hidden */
 export const joinPath = posix.join;