+++ src/util/joinPath.web.ts
@@ -1,1 +1,2 @@
+/** @hidden */
 export { join as joinPath } from "path";